% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doubleblock.R
\name{doubleblock}
\alias{doubleblock}
\title{Generating special class of (covariance) matrices}
\usage{
doubleblock(p1, p2, M, a, b)
}
\arguments{
\item{p1}{The number of x axis (latitude)}

\item{p2}{The number of y axis (longitude)}

\item{M}{Multiplicative constant for off-diagonal elements}

\item{a}{decay rate for the x axis}

\item{b}{decay rate for the y axis}
}
\value{
The block bandable covariance matrix
}
\description{
This is the function that used to generate the block bandable matrices \code{S}
with \code{p1} by \code{p2} dimensions.
}
\examples{
doubleblock(5,5,1,1.5,1.5)
}

