% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doubletaper.R
\name{doubletaper}
\alias{doubletaper}
\title{Estimator for the spatial double block matrices -- double tapering estimator}
\usage{
doubletaper(S, p1, p2, k, l, ck = 1, cl = 1, a = 1.3, b = 1.3,
  n = 1000, method = "tapering", Axis = 0)
}
\arguments{
\item{S}{The block bandable matrices}

\item{p1}{The number of x axis (latitude)}

\item{p2}{The number of y axis (longitude)}

\item{k}{bandwidth for the x axis}

\item{l}{bandwidth for the y axis}

\item{ck}{constant in the bandwidth for the x axis}

\item{cl}{constant in the bandwidth for the y axis}

\item{a}{decay rate for the x axis}

\item{b}{decay rate for the y axis}

\item{n}{sample size}

\item{method}{tapering method, "tapering" for linear tapering, "banding" for banding estimator}

\item{Axis}{tapering direction, 0 for tapering for both direction, 1 for x direction only, 2 for y direction only}
}
\value{
The regularized block bandable covariance matrix with double tapering estimator
}
\description{
This is the function that used to implement the double tapering estimator for
the block bandable matrices \code{S} with \code{p1} by \code{p2} dimensions.
}
\examples{
library(MASS)
 p1=25; p2=25; p=p1*p2; M=1; ck=cl=10; n=500;
 nalpha=1:5
 err_hat=err_band=err_taper=rep(0,length(nalpha))
 for(i in nalpha){
   a=1+i/10
   cat("Decay Rate = ",a,"...\\n")
   S=doubleblock(p1,p2,M,a,a)
   data = mvrnorm(n, mu = rep(0,p), Sigma = S)
   S_hat=cov(data)
   S_band=doubletaper(S_hat,p1,p2,ck,cl,a,a,n,method="banding",Axis=0)
   S_taper=doubletaper(S_hat,p1,p2,ck,cl,a,a,n,method="tapering",Axis=0)

   err_hat[i]=sqrt(sum((S_hat-S)^2))
   err_band[i]=sqrt(sum((S_band-S)^2))
   err_taper[i]=sqrt(sum((S_taper-S)^2))
 }
 plot(1+nalpha/10,err_hat/err_hat,col="black",type="o",ylim=c(0,1.5),
      xlab="decay rate",ylab="Relative Error",main="Relative Error for Different Estimators")
 lines(1+nalpha/10,err_band/err_hat,col="blue",type="o")
 lines(1+nalpha/10,err_taper/err_hat,col="red",type="o")
 legend("topright",c("Sample Covariance","Banding","Tapering"),col=c("black","blue","red"),
        lty=rep(1,3),pch=rep(1,3),cex=0.5)
}

